pyproject.toml
[tool.poetry]
name = "coda-ai-assistant"
version = "0.0.3"
description = "A general-purpose AI coding assistant"
authors = ["Vitaly Tsaplin"]
license = "MIT"
repository = "https://github.com/vtsaplin/coda"
homepage = "https://github.com/vtsaplin/coda"
keywords = ["AI", "coding assistant", "automation"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.10"
openai = "*"
python-dotenv = "*"
requests = "*"
requests_toolbelt = "*"
jsonschema = "*"
backoff = "*"
langchain = "*"
langchain-openai = "*"
tqdm = "*"
chromadb = "*"
colorama = "*"
unidiff = "*"
pygments = "*"
pathspec = "*"
InquirerPy = "*"
tornado = "*"
humps = "*"

[tool.poetry.scripts]
coda = "app.app:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"